Tweaked total market stuff slightly as per the comments on #167.

+++ server/static/survey/assets/js/controllers/MarketReport.js
@@ -41,7 +41,7 @@ angular.module('askApp')
             });
             console.log("Filtered: ", filtered_answers);
             $scope.total_market_weight = _.reduce(filtered_answers, function(accum, val) {
-                return accum + _.reduce(val.value, function(x,y) { return x + parseFloat(y.sum); }, 0);
+                return accum + _.reduce(val.value, function(x,y) { return x + parseInt(y.sum); }, 0);
             }, 0);
         });
     }
